requirements.txt
flask==3.0.0
librosa==0.10.1
soundfile==0.12.1
numpy>=1.24.0,<2.0.0
praat-parselmouth==0.4.3
gunicorn==21.2.0
matplotlib==3.8.2
